constants refactor
